scraper/pyproject.toml
[tool.poetry]
name = "comme-un-poisson-dans-l-eau"
version = "0.1.0"
description = ""
authors = ["Adrien Di Pasquale <adrien@dipasquale.fr>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
python-slugify = "^8.0.1"
pytube = "^15.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
